+++ code/src/cljs/sixsq/nuvla/ui/credentials/views.cljs
@@ -54,6 +54,7 @@
                          :placeholder   (@tr [name-kw])
                          :disabled      (not editable?)
                          :error         (and validate? (not valid?))
+                         :auto-complete "off"
                          :fluid         true
                          :type          (if (= type :input) :text type)
                          :icon          (when input-active? :pencil)
@@ -128,9 +129,9 @@
 
         [:<>
 
-         [acl/AclButton {:acl       (:acl @credential)
-                         :read-only (not editable?)
-                         :on-change #(dispatch [::events/update-credential :acl %])}]
+         [acl/AclButton {:default-value (:acl @credential)
+                         :read-only     (not editable?)
+                         :on-change     #(dispatch [::events/update-credential :acl %])}]
 
          [ui/Table (assoc style/definition :class :nuvla-ui-editable)
           [ui/TableBody
@@ -157,9 +158,9 @@
             form-validation-event ::events/validate-minio-credential-form]
 
         [:<>
-         [acl/AclButton {:acl       (:acl @credential)
-                         :read-only (not editable?)
-                         :on-change #(dispatch [::events/update-credential :acl %])}]
+         [acl/AclButton {:default-value (:acl @credential)
+                         :read-only     (not editable?)
+                         :on-change     #(dispatch [::events/update-credential :acl %])}]
 
          [ui/Table (assoc style/definition :class :nuvla-ui-editable)
           [ui/TableBody