+++ code/src/cljs/sixsq/nuvla/ui/history/views.cljs
@@ -12,9 +12,9 @@
    value will also be used as the label, unless an explicit label is provided."
   [href & [label]]
   [:a {:href     (str @config/path-prefix "/" href)
-       :style {:overflow      "hidden",
-               :text-overflow "ellipsis",
-               :max-width     "20ch"}
+       :style    {:overflow      "hidden",
+                  :text-overflow "ellipsis",
+                  :max-width     "20ch"}
        :target   "_blank"
        :on-click (fn [event]
                    (when-not (.-metaKey event)              ;;cmd key not pressed